feat: better use of the modal + coderabbit comments
