+++ packages/common-config/src/types/docusaurus__link.d.ts
@@ -1,7 +1,8 @@
 declare module '@docusaurus/Link' {
   import * as React from 'react';
   interface LinkProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
-    to: string;
+    to?: string;
+    href?: string;
     className?: string;
     children?: React.ReactNode;
   }
