+++ sites/template/docusaurus.config.ts
@@ -62,6 +62,12 @@ const localConfig: Config = {
           position: "right",
           label: "Website",
         },
+        {
+          type: "html",
+          position: "right",
+          value:
+            '<style>#disclaimer-btn::before { content: ""; width: 16px; height: 16px; background-image: url("/common/img/icons/information-circle.svg"); background-size: contain; background-repeat: no-repeat; margin-right: 8px; } #disclaimer-btn:hover { color: #d1d5db; } @media (max-width: 996px) { #disclaimer-btn { display: none !important; } }</style><button id="disclaimer-btn" style="background: none; border: none; color: #5D6B98; cursor: pointer; margin-left: 8px; display: flex; align-items: center; font-size: 14px; font-family: inherit; transition: color 0.2s ease;">Disclaimer</button>',
+        },
       ],
     },
   } satisfies Preset.ThemeConfig,
